+++ History.md
@@ -1,5 +1,9 @@
 ## v.NEXT
 
+* An `onExternalLogin` hook has been added to the accounts system, to allow
+  the customization of OAuth user profile updates.
+  [PR #9042](https://github.com/meteor/meteor/pull/9042)
+
 * iOS icons and launch screens have been updated to support iOS 11
   [Issue #9196](https://github.com/meteor/meteor/issues/9196)
   [PR #9198](https://github.com/meteor/meteor/pull/9198)